refactor: ui 폴더 개선
