+++ src/app/App.tsx
@@ -1,5 +1,5 @@
 import PostManagerPage from "@pages/PostsManagerPage"
-import Layout from "@shared/ui/Layout"
+import Layout from "@shared/ui/Layout/Layout"
 import { BrowserRouter as Router } from "react-router-dom"
 
 const App = () => {
