+++ src/shared/ui/index.tsx
@@ -3,5 +3,5 @@ export { Card, CardTitle, CardHeader, CardContent } from "./Card"
 export { Select, SelectGroup, SelectValue, SelectTrigger, SelectContent, SelectItem } from "./Select"
 export { Dialog, DialogTrigger, DialogPortal, DialogOverlay, DialogContent, DialogHeader, DialogTitle } from "./Dialog"
 export { Table, TableBody, TableCell, TableHead, TableHeader, TableRow } from "./Table"
-export { Input } from "./Input"
-export { Textarea } from "./TextArea"
+export { Input, Textarea } from "./TextField"
+export { Pagination } from "./Pagination"
